requirements.txt
langchain
langchain-community
streamlit
supabase
PyPDF2
faiss-cpu
sentence-transformers
google-generativeai
python-dotenv